ref: Move part of path code into utils
