+++ src/actors/objects/s3.rs
@@ -10,7 +10,7 @@ use tokio::codec::{BytesCodec, FramedRead};
 use tokio_threadpool::ThreadPool;
 
 use crate::actors::common::cache::Cacher;
-use crate::actors::objects::paths::prepare_download_paths;
+use crate::actors::objects::common::prepare_download_paths;
 use crate::actors::objects::{
     DownloadPath, DownloadStream, FetchFileInner, FetchFileRequest, ObjectError, ObjectErrorKind,
     PrioritizedDownloads,
